+++ packages/cortex-react/src/components/Select/Trigger.tsx
@@ -1,10 +1,10 @@
-import { labelStyle, selectVariants } from '@tecsinapse/cortex-core';
-import React, { useContext, useMemo } from 'react';
+import React, { useContext, useEffect, useRef, useMemo } from 'react';
 import { IoChevronDownOutline } from 'react-icons/io5';
 import { Popover } from '../Popover';
 import { SelectContext } from './context';
 import { SelectTriggerProps } from './types';
 import clsx from 'clsx';
+import { labelStyle, selectVariants } from '@tecsinapse/cortex-core';
 
 const { button } = selectVariants();
 
@@ -15,7 +15,9 @@ export const SelectTrigger = ({
   multiLabelSelected,
   ...rest
 }: SelectTriggerProps) => {
-  const { value, labelExtractor } = useContext(SelectContext);
+  const { value, labelExtractor, setTriggerWidth } = useContext(SelectContext);
+  const triggerRef = useRef<HTMLButtonElement>(null);
+
   const _placeholder = useMemo(() => {
     if (value?.length === 0 || !value) return placeholder ?? label;
     if (value?.length === 1) return labelExtractor(value[0]);
@@ -26,14 +28,21 @@ export const SelectTrigger = ({
     return labelExtractor(value);
   }, [label, value]);
 
-  const { className } = rest;
-
   const hasValue = Array.isArray(value) ? value.length > 0 : !!value;
+
+  useEffect(() => {
+    if (triggerRef.current && setTriggerWidth) {
+      const width = triggerRef.current.getBoundingClientRect().width;
+      setTriggerWidth(width);
+    }
+  }, [triggerRef.current, setTriggerWidth]);
+
   return (
     <Popover.Trigger disabled={disabled}>
       <div className="w-full gap-mili">
         <button
-          className={button({ className })}
+          ref={triggerRef}
+          className={button({ className: rest.className })}
           disabled={disabled}
           role="button"
           {...rest}
@@ -57,9 +66,7 @@ export const SelectTrigger = ({
           >
             {label}
           </label>
-        ) : (
-          <></>
-        )}
+        ) : null}
       </div>
     </Popover.Trigger>
   );